refactor(networks): modularize wallet network config
