+++ networks/cardano/network-cardano-suite-common/package.json
@@ -13,6 +13,7 @@
     },
     "dependencies": {
         "@scure/base": "^2.0.0",
+        "@trezor/device-utils": "workspace:*",
         "@trezor/network-cardano": "workspace:*",
         "@trezor/network-module-suite-common-types": "workspace:*",
         "cbor": "^10.0.12",
