+++ networks/ethereum/network-ethereum-suite-common/src/networkConfig.ts
@@ -1,18 +1,133 @@
 import type { EthereumNetworkSymbol } from '@trezor/network-ethereum/constants';
 import {
+    type Explorer,
     type SuiteCommonNetworkConfig,
     asProtocol,
 } from '@trezor/network-module-suite-common-types';
 
+const getExplorerUrls = (baseUrl: string): Explorer => ({
+    base: baseUrl,
+    tx: `${baseUrl}/tx/`,
+    address: `${baseUrl}/address/`,
+    nft: `${baseUrl}/nft/`,
+});
+
 const networkConfigBySymbol: Readonly<Record<EthereumNetworkSymbol, SuiteCommonNetworkConfig>> = {
-    eth: { color: '#454a75', protocols: [asProtocol('ethereum'), asProtocol('eth')] },
+    eth: {
+        color: '#454a75',
+        protocols: [asProtocol('ethereum'), asProtocol('eth')],
+        displaySymbol: 'ETH',
+        name: 'Ethereum',
+        networkType: 'ethereum',
+        chainId: 1,
+        bip43Path: "m/44'/60'/0'/0/i",
+        decimals: 18,
+        testnet: false,
+        explorer: getExplorerUrls('https://etherscan.io'),
+        features: [
+            'rbf',
+            'sign-verify',
+            'tokens',
+            'nfts',
+            'coin-definitions',
+            'nft-definitions',
+            'staking',
+            'eip1559',
+            'mev-protection',
+            'graph',
+            'claim-rewards',
+        ],
+        backendOptions: [{ type: 'blockbook' }, { type: 'evm-rpc' }],
+        accountTypes: {
+            ledger: {
+                // ledger (live), #1 acc is same as Trezor, so it is skipped
+                accountType: 'ledger',
+                bip43Path: "m/44'/60'/i'/0/0",
+                isDebugOnlyAccountType: true,
+            },
+            legacy: {
+                // ledger (legacy)
+                accountType: 'legacy',
+                bip43Path: "m/44'/60'/0'/i",
+                isDebugOnlyAccountType: true,
+            },
+        },
+        coingeckoId: 'ethereum',
+        tradeCryptoId: 'ethereum',
+        caipId: 'eip155:1',
+        yieldXyzId: 'ethereum',
+    },
     pol: {
         color: '#7b3fe4',
         protocols: [asProtocol('polygon'), asProtocol('matic'), asProtocol('pol')],
+        displaySymbol: 'POL',
+        displaySymbolName: 'Polygon',
+        name: 'Polygon PoS',
+        networkType: 'ethereum',
+        chainId: 137,
+        bip43Path: "m/44'/60'/0'/0/i",
+        decimals: 18,
+        testnet: false,
+        explorer: getExplorerUrls('https://polygonscan.com'),
+        features: [
+            'rbf',
+            'sign-verify',
+            'tokens',
+            'nfts',
+            'coin-definitions',
+            'nft-definitions',
+            'eip1559',
+            'graph',
+        ],
+        backendOptions: [{ type: 'blockbook', isExternalBackend: true }, { type: 'evm-rpc' }],
+        accountTypes: {
+            ledger: {
+                // ledger (live), #1 acc is same as Trezor, so it is skipped
+                accountType: 'ledger',
+                bip43Path: "m/44'/60'/i'/0/0",
+                isDebugOnlyAccountType: true,
+            },
+        },
+        coingeckoId: 'polygon-pos',
+        tradeCryptoId: 'polygon-ecosystem-token',
+        caipId: 'eip155:137',
+        yieldXyzId: 'polygon',
     },
     bsc: {
         color: '#f0b90b',
         protocols: [asProtocol('binance'), asProtocol('bnb'), asProtocol('bsc')],
+        displaySymbol: 'BNB',
+        displaySymbolName: 'BNB',
+        name: 'BNB Smart Chain',
+        networkType: 'ethereum',
+        chainId: 56,
+        bip43Path: "m/44'/60'/0'/0/i",
+        decimals: 18,
+        testnet: false,
+        explorer: getExplorerUrls('https://bscscan.com'),
+        features: [
+            'rbf',
+            'sign-verify',
+            'tokens',
+            'nfts',
+            'coin-definitions',
+            'nft-definitions',
+            'mev-protection',
+            'graph',
+        ],
+        backendOptions: [{ type: 'blockbook', isExternalBackend: true }, { type: 'evm-rpc' }],
+        accountTypes: {
+            ledger: {
+                // ledger (live), #1 acc is same as Trezor, so it is skipped
+                accountType: 'ledger',
+                bip43Path: "m/44'/60'/i'/0/0",
+                isDebugOnlyAccountType: true,
+            },
+        },
+        coingeckoId: 'binance-smart-chain',
+        tradeCryptoId: 'binancecoin',
+        caipId: 'eip155:56',
+        yieldXyzId: 'binance',
     },
     arb: {
         color: '#213147',
@@ -22,21 +137,286 @@ const networkConfigBySymbol: Readonly<Record<EthereumNetworkSymbol, SuiteCommonN
             asProtocol('arb'),
             asProtocol('arbitrum-ethereum'),
         ],
+        settlementLayer: 'eth',
+        displaySymbol: 'ETH',
+        displaySymbolName: 'Arbitrum One Ethereum',
+        name: 'Arbitrum One',
+        networkType: 'ethereum',
+        chainId: 42161,
+        bip43Path: "m/44'/60'/0'/0/i",
+        decimals: 18,
+        testnet: false,
+        explorer: getExplorerUrls('https://arbiscan.io'),
+        features: [
+            'rbf',
+            'sign-verify',
+            'tokens',
+            'nfts',
+            'coin-definitions',
+            'nft-definitions',
+            'mev-protection',
+            'eip1559',
+            'graph',
+        ],
+        backendOptions: [{ type: 'blockbook', isExternalBackend: true }, { type: 'evm-rpc' }],
+        accountTypes: {
+            ledger: {
+                // ledger (live), #1 acc is same as Trezor, so it is skipped
+                accountType: 'ledger',
+                bip43Path: "m/44'/60'/i'/0/0",
+                isDebugOnlyAccountType: true,
+            },
+        },
+        coingeckoId: 'arbitrum-one',
+        tradeCryptoId: 'arbitrum-one--0x0000000000000000000000000000000000000000',
+        caipId: 'eip155:42161',
+        yieldXyzId: 'arbitrum',
     },
-    base: { color: '#0052ff', protocols: [asProtocol('base')] },
-    op: { color: '#ff0720', protocols: [asProtocol('optimism'), asProtocol('op')] },
+    base: {
+        color: '#0052ff',
+        protocols: [asProtocol('base')],
+        settlementLayer: 'eth',
+        displaySymbol: 'ETH',
+        displaySymbolName: 'Base Ethereum',
+        name: 'Base',
+        networkType: 'ethereum',
+        chainId: 8453,
+        bip43Path: "m/44'/60'/0'/0/i",
+        decimals: 18,
+        testnet: false,
+        explorer: getExplorerUrls('https://basescan.org'),
+        features: [
+            'rbf',
+            'sign-verify',
+            'tokens',
+            'nfts',
+            'coin-definitions',
+            'nft-definitions',
+            'eip1559',
+            'mev-protection',
+            'graph',
+        ],
+        backendOptions: [{ type: 'blockbook', isExternalBackend: true }, { type: 'evm-rpc' }],
+        accountTypes: {
+            ledger: {
+                // ledger (live), #1 acc is same as Trezor, so it is skipped
+                accountType: 'ledger',
+                bip43Path: "m/44'/60'/i'/0/0",
+                isDebugOnlyAccountType: true,
+            },
+        },
+        coingeckoId: 'base',
+        tradeCryptoId: 'base--0x0000000000000000000000000000000000000000',
+        caipId: 'eip155:8453',
+        nativeTokenReserve: '0.0002',
+        yieldXyzId: 'base',
+    },
+    op: {
+        color: '#ff0720',
+        protocols: [asProtocol('optimism'), asProtocol('op')],
+        settlementLayer: 'eth',
+        displaySymbol: 'ETH',
+        displaySymbolName: 'Optimism Ethereum',
+        name: 'Optimism',
+        networkType: 'ethereum',
+        chainId: 10,
+        bip43Path: "m/44'/60'/0'/0/i",
+        decimals: 18,
+        testnet: false,
+        explorer: getExplorerUrls('https://optimistic.etherscan.io'),
+        features: [
+            'rbf',
+            'sign-verify',
+            'tokens',
+            'nfts',
+            'coin-definitions',
+            'nft-definitions',
+            'eip1559',
+            'graph',
+        ],
+        backendOptions: [{ type: 'blockbook', isExternalBackend: true }, { type: 'evm-rpc' }],
+        accountTypes: {
+            ledger: {
+                // ledger (live), #1 acc is same as Trezor, so it is skipped
+                accountType: 'ledger',
+                bip43Path: "m/44'/60'/i'/0/0",
+                isDebugOnlyAccountType: true,
+            },
+        },
+        coingeckoId: 'optimistic-ethereum',
+        tradeCryptoId: 'optimistic-ethereum--0x0000000000000000000000000000000000000000',
+        caipId: 'eip155:10',
+        nativeTokenReserve: '0.0002',
+        yieldXyzId: 'optimism',
+    },
     rhc: {
         color: '#ccff00',
         protocols: [asProtocol('robinhood'), asProtocol('robinhood-chain'), asProtocol('rhc')],
+        settlementLayer: 'eth',
+        displaySymbol: 'ETH',
+        displaySymbolName: 'Robinhood Ethereum',
+        name: 'Robinhood Chain',
+        networkType: 'ethereum',
+        chainId: 4663,
+        bip43Path: "m/44'/60'/0'/0/i",
+        decimals: 18,
+        testnet: false,
+        explorer: getExplorerUrls('https://robinscan.io'),
+        features: ['rbf', 'sign-verify', 'tokens', 'nfts', 'coin-definitions', 'eip1559', 'graph'],
+        backendOptions: [{ type: 'blockbook', isExternalBackend: true }, { type: 'evm-rpc' }],
+        accountTypes: {
+            ledger: {
+                // ledger (live), #1 acc is same as Trezor, so it is skipped
+                accountType: 'ledger',
+                bip43Path: "m/44'/60'/i'/0/0",
+                isDebugOnlyAccountType: true,
+            },
+        },
+        coingeckoId: 'robinhood',
+        tradeCryptoId: 'robinhood--0x0000000000000000000000000000000000000000',
+        caipId: 'eip155:4663',
+        nativeTokenReserve: '0.0002',
+        yieldXyzId: null,
     },
     hype: {
         color: '#97fce4',
         protocols: [asProtocol('hyperliquid'), asProtocol('hyperevm'), asProtocol('hype')],
+        displaySymbol: 'HYPE',
+        name: 'HyperEVM',
+        displaySymbolName: 'Hyperliquid',
+        networkType: 'ethereum',
+        chainId: 999,
+        bip43Path: "m/44'/60'/0'/0/i",
+        decimals: 18,
+        testnet: false,
+        explorer: getExplorerUrls('https://hyperevmscan.io'),
+        features: [
+            'rbf',
+            'sign-verify',
+            'tokens',
+            'nfts',
+            'coin-definitions',
+            'nft-definitions',
+            'eip1559',
+            'graph',
+        ],
+        backendOptions: [{ type: 'blockbook', isExternalBackend: true }, { type: 'evm-rpc' }],
+        accountTypes: {
+            ledger: {
+                // ledger (live), #1 acc is same as Trezor, so it is skipped
+                accountType: 'ledger',
+                bip43Path: "m/44'/60'/i'/0/0",
+                isDebugOnlyAccountType: true,
+            },
+        },
+        coingeckoId: 'hyperevm',
+        tradeCryptoId: 'hyperevm',
+        caipId: 'eip155:999',
+        isDebugOnlyNetwork: true,
+        yieldXyzId: null,
     },
-    avax: { color: '#e84142', protocols: [asProtocol('avalanche'), asProtocol('avax')] },
-    etc: { color: '#60c67e', protocols: [asProtocol('ethclassic'), asProtocol('etc')] },
-    tsep: { color: '#454a75', protocols: [asProtocol('tsep')] },
-    thod: { color: '#454a75', protocols: [asProtocol('thod')] },
+    avax: {
+        color: '#e84142',
+        protocols: [asProtocol('avalanche'), asProtocol('avax')],
+        displaySymbol: 'AVAX',
+        displaySymbolName: 'Avalanche',
+        name: 'Avalanche C-Chain',
+        networkType: 'ethereum',
+        chainId: 43114,
+        bip43Path: "m/44'/60'/0'/0/i",
+        decimals: 18,
+        testnet: false,
+        explorer: getExplorerUrls('https://snowscan.xyz'),
+        features: [
+            'rbf',
+            'sign-verify',
+            'tokens',
+            'nfts',
+            'coin-definitions',
+            'nft-definitions',
+            'eip1559',
+            'graph',
+        ],
+        backendOptions: [{ type: 'blockbook', isExternalBackend: true }, { type: 'evm-rpc' }],
+        accountTypes: {
+            ledger: {
+                // ledger (live), #1 acc is same as Trezor, so it is skipped
+                accountType: 'ledger',
+                bip43Path: "m/44'/60'/i'/0/0",
+                isDebugOnlyAccountType: true,
+            },
+        },
+        coingeckoId: 'avalanche',
+        tradeCryptoId: 'avalanche-2',
+        caipId: 'eip155:43114',
+        yieldXyzId: 'avalanche-c',
+    },
+    etc: {
+        color: '#60c67e',
+        protocols: [asProtocol('ethclassic'), asProtocol('etc')],
+        displaySymbol: 'ETC',
+        name: 'Ethereum Classic',
+        networkType: 'ethereum',
+        chainId: 61,
+        bip43Path: "m/44'/61'/0'/0/i",
+        decimals: 18,
+        testnet: false,
+        explorer: getExplorerUrls('https://etc.trezor.io'),
+        features: ['sign-verify', 'tokens', 'coin-definitions', 'graph'],
+        backendOptions: [{ type: 'blockbook' }, { type: 'evm-rpc' }],
+        accountTypes: {},
+        coingeckoId: 'ethereum-classic',
+        tradeCryptoId: 'ethereum-classic',
+        yieldXyzId: null,
+    },
+    tsep: {
+        color: '#454a75',
+        protocols: [asProtocol('tsep')],
+        displaySymbol: 'tETH',
+        name: 'Ethereum Sepolia',
+        networkType: 'ethereum',
+        bip43Path: "m/44'/60'/0'/0/i",
+        chainId: 11155111,
+        decimals: 18,
+        testnet: true,
+        explorer: getExplorerUrls('https://sepolia.etherscan.io'),
+        features: ['rbf', 'sign-verify', 'tokens', 'nfts', 'eip1559', 'graph'],
+        backendOptions: [{ type: 'blockbook' }, { type: 'evm-rpc' }],
+        accountTypes: {
+            legacy: {
+                accountType: 'legacy',
+                bip43Path: "m/44'/1'/0'/0/i",
+                isDebugOnlyAccountType: true,
+            },
+        },
+        coingeckoId: 'sepolia-test-ethereum', // fake, coingecko does not have testnets
+        tradeCryptoId: 'sepolia-test-ethereum', // fake, coingecko does not have testnets
+        yieldXyzId: 'ethereum-sepolia',
+    },
+    thod: {
+        color: '#454a75',
+        protocols: [asProtocol('thod')],
+        displaySymbol: 'tETH',
+        name: 'Ethereum Hoodi',
+        networkType: 'ethereum',
+        bip43Path: "m/44'/60'/0'/0/i",
+        chainId: 560048,
+        decimals: 18,
+        testnet: true,
+        explorer: getExplorerUrls('https://hoodi.etherscan.io'),
+        features: ['rbf', 'sign-verify', 'tokens', 'staking', 'nfts', 'eip1559', 'graph'],
+        backendOptions: [{ type: 'blockbook' }, { type: 'evm-rpc' }],
+        accountTypes: {
+            legacy: {
+                accountType: 'legacy',
+                bip43Path: "m/44'/1'/0'/0/i",
+                isDebugOnlyAccountType: true,
+            },
+        },
+        coingeckoId: 'hoodi-test-ethereum', // fake, coingecko does not have testnets
+        tradeCryptoId: 'hoodi-test-ethereum', // fake, coingecko does not have testnets
+        yieldXyzId: 'ethereum-hoodi',
+    },
 };
 
 export const getNetworkConfig = (symbol: EthereumNetworkSymbol): SuiteCommonNetworkConfig =>
