+++ networks/network-module/network-module-suite-common-types/package.json
@@ -11,6 +11,8 @@
         "type-check": "yarn g:tsc --build"
     },
     "dependencies": {
+        "@trezor/crypto-utils": "workspace:*",
+        "@trezor/device-utils": "workspace:*",
         "@trezor/type-utils": "workspace:*"
     }
 }
