+++ networks/network-module/network-module-suite-common-types/src/SuiteCommonNetworkModule.ts
@@ -1,12 +1,5 @@
 import type { AddressValidator } from './AddressValidator';
-import type { Protocol } from './Protocol';
-
-export type NetworkColor = `#${string}`;
-
-export type SuiteCommonNetworkConfig = {
-    readonly color: NetworkColor;
-    readonly protocols: readonly Protocol[];
-};
+import type { SuiteCommonNetworkConfig } from './SuiteCommonNetworkConfig';
 
 export type SuiteCommonNetworkModule<TSymbol extends string> = {
     addressValidator: AddressValidator<TSymbol>;
