+++ networks/network-module/network-module-suite-common-types/src/index.ts
@@ -2,8 +2,17 @@ export { addressType } from './AddressValidator';
 export type { AddressType, AddressValidator } from './AddressValidator';
 export { asProtocol } from './Protocol';
 export type { Protocol } from './Protocol';
-export type {
-    NetworkColor,
-    SuiteCommonNetworkConfig,
-    SuiteCommonNetworkModule,
-} from './SuiteCommonNetworkModule';
+export type { SuiteCommonNetworkModule } from './SuiteCommonNetworkModule';
+export {
+    TREZOR_CONNECT_BACKENDS,
+    type AccountType,
+    type BackendType,
+    type Explorer,
+    type NetworkAccount,
+    type NetworkColor,
+    type NetworkFeature,
+    type NetworkType,
+    type ServerType,
+    type SuiteCommonNetworkConfig,
+    type TrezorConnectBackendType,
+} from './SuiteCommonNetworkConfig';
