+++ networks/network-module/network-module-suite-common-types/tsconfig.json
@@ -2,6 +2,12 @@
     "extends": "../../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
     "references": [
+        {
+            "path": "../../../packages/crypto-utils"
+        },
+        {
+            "path": "../../../packages/device-utils"
+        },
         { "path": "../../../packages/type-utils" }
     ]
 }
