+++ networks/network-module/network-module-suite-common-types/tsconfig.lib.json
@@ -6,6 +6,14 @@
     },
     "include": ["./src"],
     "references": [
-        { "path": "../../../packages/type-utils" }
+        {
+            "path": "../../../packages/crypto-utils"
+        },
+        {
+            "path": "../../../packages/device-utils"
+        },
+        {
+            "path": "../../../packages/type-utils"
+        }
     ]
 }
