+++ networks/solana/network-solana-suite-common/package.json
@@ -13,6 +13,7 @@
     },
     "dependencies": {
         "@scure/base": "^2.0.0",
+        "@trezor/device-utils": "workspace:*",
         "@trezor/network-module-suite-common-types": "workspace:*",
         "@trezor/network-solana": "workspace:*"
     }
