+++ networks/tron/network-tron-suite-common/package.json
@@ -14,6 +14,7 @@
     "dependencies": {
         "@noble/hashes": "^2.0.1",
         "@scure/base": "^2.0.0",
+        "@trezor/device-utils": "workspace:*",
         "@trezor/network-module-suite-common-types": "workspace:*",
         "@trezor/network-tron": "workspace:*"
     }
