+++ suite-common/wallet-config/src/index.ts
@@ -1,8 +1,71 @@
-export * from './backends';
-export * from './earnRewardsProvider';
-export * from './networksConfig';
-export * from './representativeAssets';
-export * from './stakingProviders';
-export * from './types';
-export * from './utils';
-export * from './getExplorerUrls';
+export { getServerAddressExample, validateServerAddress } from './backends';
+export {
+    EARN_YIELD_CLAIM_PROVIDER,
+    getEarnYieldClaimContractAddress,
+    isEarnYieldClaimSupported,
+} from './earnRewardsProvider';
+export {
+    PROD_STAKING_SYMBOLS,
+    STAKING_SYMBOLS,
+    STAKING_TYPES,
+    networks,
+    type StakingNetworkSymbol,
+    type StakingNetworkType,
+} from './networksConfig';
+export { getRepresentativeAssets, type RepresentativeAsset } from './representativeAssets';
+export {
+    getStakingProviderByCardanoPoolId,
+    getStakingProviderByEthereumPoolName,
+    getStakingProviderBySolanaVoterPubkey,
+    getStakingProviderByTronSrAddress,
+    type StakingProvider,
+} from './stakingProviders';
+export {
+    TREZOR_CONNECT_BACKENDS,
+    asNetworkSymbol,
+    toNetworkSymbolNonTestnet,
+    type AccountType,
+    type BackendType,
+    type Explorer,
+    type Network,
+    type NetworkAccount,
+    type NetworkDisplaySymbol,
+    type NetworkFeature,
+    type NetworkSymbol,
+    type NetworkSymbolExtended,
+    type NetworkType,
+    type Networks,
+    type ServerType,
+    type TrezorConnectBackendType,
+} from './types';
+export {
+    NORMAL_ACCOUNT_TYPE,
+    filterNetworksByName,
+    getCoingeckoId,
+    getDisplaySymbol,
+    getMainnets,
+    getNetwork,
+    getNetworkByCoingeckoId,
+    getNetworkByEvmChainId,
+    getNetworkByTradeCryptoId,
+    getNetworkByYieldXyzId,
+    getNetworkChainId,
+    getNetworkDecimals,
+    getNetworkDisplaySymbol,
+    getNetworkDisplaySymbolName,
+    getNetworkFeatures,
+    getNetworkOptional,
+    getNetworkType,
+    getNetworksWithMevProtection,
+    getNetworksWithNativeTokenReserve,
+    getTestnetSymbols,
+    getTestnets,
+    isAccountBasedNetwork,
+    isAccountOfNetwork,
+    isBlockbookBasedNetwork,
+    isNetworkSymbol,
+    isNetworkUsingExternalBackend,
+    networkSymbolCollection,
+    networksCollection,
+} from './utils';
+export { getExplorerUrl, getParsedExplorerUrls } from './getExplorerUrls';
