+++ suite-common/wallet-utils/src/reviewTransactionUtils.ts
@@ -164,7 +164,7 @@ export const getClearSignedEvmTradingSwapCoverage = ({
         return undefined;
     }
     const network = networks[account.symbol];
-    if (!('chainId' in network)) {
+    if (network.chainId === undefined) {
         return undefined;
     }
     const to = precomposedTx.outputs.find(
