+++ suite/e2e/support/types/index.ts
@@ -2,7 +2,7 @@ import type { CryptoId } from 'invity-api';
 import { RequireExactlyOne } from 'type-fest';
 
 import { AnalyticsDesktopEvents } from '@suite/analytics';
-import type { NetworkConfigWithoutTestnets, NetworkSymbol } from '@suite-common/wallet-config';
+import type { NetworkSymbol } from '@suite-common/wallet-config';
 import type { TrezorUserEnvLinkClass } from '@trezor/trezor-user-env-link';
 
 import { LaunchSuiteParams } from '../electron';
@@ -62,7 +62,7 @@ export type TrezorUserEnv = Pick<
     | 'sendToAddressAndMineBlock'
 >;
 
-export type AssetPickerNetworkFilter = 'all-networks' | NetworkConfigWithoutTestnets['symbol'];
+export type AssetPickerNetworkFilter = 'all-networks' | NetworkSymbol;
 
 export type BuyAsset = RequireExactlyOne<
     {
